override built-in date pipe with i18n-ready custom pipe
Angular's built-in date pipe is overridden with a custom pipe that does
essentially the same thing (it purely forwards the inputs to the
built-in pipe, therefore it's called DateProxyPipe), BUT is able to take
language into account EVEN when the language is changed during runtime.

`registerLocaleData` is necessary to load information on what e.g. the
"medium" date format should look like in German etc.

The HelgolandPipesModule must be imported in any sub-module that should
have the built-in date pipe overwritten.